+++ CHANGELOG.md
@@ -1,3 +1,4 @@
+# 2025-05-05 Version 0.10
 
 - Add `Thread#machine`
 - Add `Fiber#mailbox`